andando a medias el endpoint de ver detalle de proyecto

+++ dssd-backend/src/main/java/grupo16/dssd_backend/repositories/ProyectoRepository.java
@@ -9,5 +9,6 @@ import java.util.List;
 @Repository
 public interface ProyectoRepository extends JpaRepository<Proyecto, Long> {
 
-    List<Proyecto> findByCaseIdIn(List<Integer> caseIds);
+    List<Proyecto> findByCaseIdIn(List<Long> caseIds);
+
 }
